requirements.txt
torch
transformers
pdf2image
Pillow
